requirements.txt
python-telegram-bot==13.7
requests==2.26.0
sqlite3==2.6.0
python-dotenv==0.19.0
Flask==2.0.1